PA_simulator/requirements.txt
action-tutorials-py==0.14.4
actionlib==1.14.3
ament-cmake-test==1.1.6
ament-copyright==0.10.7
ament-cppcheck==0.10.7
ament-cpplint==0.10.7
ament-flake8==0.10.7
ament-index-python==1.2.0
ament-lint==0.10.7
ament-lint-cmake==0.10.7
ament-package==0.13.1
ament-pep257==0.10.7
ament-uncrustify==0.10.7
ament-xmllint==0.10.7
angles==1.9.14
base_local_planner==1.17.3
beartype==0.16.4
bondpy==1.8.7
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
cloudpickle==3.1.1
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.1.2
deepdiff==6.7.1
defusedxml==0.7.1
demo-nodes-py==0.14.4
diagnostic-updater==1.12.1
distro==1.9.0
docutils==0.20.1
domain-coordinator==0.9.2
dynamic-reconfigure==1.7.6
examples-rclpy-executors==0.11.2
examples-rclpy-minimal-action-client==0.11.2
examples-rclpy-minimal-action-server==0.11.2
examples-rclpy-minimal-client==0.11.2
examples-rclpy-minimal-publisher==0.11.2
examples-rclpy-minimal-service==0.11.2
examples-rclpy-minimal-subscriber==0.11.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
gym-notices==0.0.8
h5py==3.10.0
image-geometry==3.1.3
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
interactive-markers==1.12.2
jedi==0.19.2
Jinja2==3.1.6
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
k-Wave-python==0.3.1
kdl-parser-py==1.14.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.9
laser-geometry==1.6.8
launch==0.17.2
launch-param-builder==0.1.1
launch-ros==0.14.5
launch-testing==0.17.2
launch-testing-ros==0.14.5
launch-xml==0.17.2
launch-yaml==0.17.2
lazy_loader==0.4
MarkupSafe==2.1.5
math-function==0.0.1
math3d==4.0.0
matlab==0.1
matlabengine==9.13.1
matplotlib==3.7.2
mbf_abstract_nav==0.4.0
message-filters==1.17.4
mkl-service==2.4.0
moveit-commander==1.1.16
moveit-configs-utils==2.3.0
moveit-core==1.1.16
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
moveit-task-constructor-core==0.1.3
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
nptyping==2.5.0
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
ordered-set==4.1.0
osrf-pycommon==2.0.0
packaging==25.0
pandas==2.0.3
parso==0.8.4
pcl==0.0.0.post1
Pillow==9.3.0
pmcx==0.4.2
pudb==2025.1
pyem==2.1.0
Pygments==2.19.2
pyparsing==3.0.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.15.0
pyros-genmsg==0.5.8
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
quality-of-service-demo-py==0.14.4
rclpy==1.9.3
rcutils==4.0.4
resource-retriever==1.12.10
ros2action==0.13.5
ros2bag==0.9.2
ros2cli==0.13.5
ros2component==0.13.5
ros2doctor==0.13.5
ros2interface==0.13.5
ros2launch==0.14.5
ros2lifecycle==0.13.5
ros2multicast==0.13.5
ros2node==0.13.5
ros2param==0.13.5
ros2pkg==0.13.5
ros2run==0.13.5
ros2service==0.13.5
ros2topic==0.13.5
rosbag==1.17.4
rosbag2-py==0.9.2
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosgraph==1.17.4
rosidl-adapter==2.2.3
rosidl-cli==2.2.3
rosidl-cmake==2.2.3
rosidl-generator-c==2.2.3
rosidl-generator-cpp==2.2.3
rosidl-generator-py==0.11.4
rosidl-parser==2.2.3
rosidl-typesupport-c==1.2.1
rosidl-typesupport-cpp==1.2.1
rosidl-typesupport-fastrtps-c==1.2.1
rosidl-typesupport-fastrtps-cpp==1.2.1
rosidl-typesupport-introspection-c==2.2.3
rosidl-typesupport-introspection-cpp==2.2.3
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rpyutils==0.2.0
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-msg==0.4.12
rqt-plot==0.4.16
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-steering==0.5.14
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rviz==1.14.26
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
sensor-msgs==1.13.2
six==1.17.0
smclib==1.8.7
srdfdom==0.6.4
sros2==0.10.3
sympy==1.13.3
teleop-twist-keyboard==2.3.2
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
tf2-ros-py==0.17.5
tf2-tools==0.17.5
tifffile==2023.7.10
topic-monitor==0.14.4
topic-tools==1.17.4
torch==1.8.2
torchaudio==0.8.2
torchvision==0.19.1
tqdm==4.67.1
trac_ik_python==1.6.7
triton==2.3.0
typing_extensions==4.13.2
urdfdom-py==0.4.6
urwid==2.6.16
urwid_readline==0.15.1
urx==0.11.0
vtk==9.5.0
wcwidth==0.2.13
xacro==1.14.20
xlwt==1.3.0
zipp==3.20.2
